CardDealer/requirements.txt
pillow==9.4.0
numpy==1.24.2
twitchio==2.9.1
requests==2.28.1
PyYAML==6.0
